Abandoned the idea of a pure Python solution. Added bridge.
This commit adds a new Python-friendly FUSE wrapper that can be
compiled as a shared library and bound with ctypes to make simple
FUSE filesystems.

This change was made for simplicity's sake. It turns out that it's
pretty non-trivial to get Python to parse all of the required
typedefs, padding, structs, #defines, unions, etc.

The resulting solution would have taken long enough to run that the
results would need to be cached somewhere, and a C compiler would be
required anyways. And in that case, why not just wrap the FUSE
bindings in C?